@@ -2,40 +2,67 @@
   "name": "texture-explorer",
   "version": "0.1.0",
   "main": "main.js",
+  "scripts": {
+    "start": "electron main.js debug",
+    "browser": "gulp"
+  },
   "dependencies": {
-    "babel": "^5.8.34",
-    "immutable": "^3.7.5",
-    "lodash": "^3.10.1",
+    "babel-core": "^6.26.0",
+    "babel-plugin-add-module-exports": "^0.2.1",
+    "babel-plugin-transform-class-properties": "^6.24.1",
+    "babel-plugin-transform-es2015-classes": "^6.24.1",
+    "babel-polyfill": "^6.26.0",
+    "babel-preset-env": "^1.6.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-stage-0": "^6.24.1",
+    "babel-register": "^6.26.0",
+    "electron-window-state": "^4.1.1",
+    "file-saver": "^1.3.3",
+    "immutable": "^3.8.1",
+    "lodash": "^4.17.4",
     "minimist": "^1.2.0",
-    "node-sass": "^3.4.2",
-    "png-stream": "^1.0.4",
-    "react": "^0.14.2",
-    "react-dom": "^0.14.2",
-    "reflux": "^0.3.0",
-    "xml2js": "^0.4.15"
+    "node-sass": "^4.5.3",
+    "png-stream": "^1.0.5",
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0",
+    "react-redux": "^5.0.6",
+    "redux": "^3.7.2",
+    "redux-thunk": "^2.2.0",
+    "webworkify": "^1.4.0",
+    "xml2js": "^0.4.19"
   },
   "devDependencies": {
-    "aliasify": "~1.7.2",
-    "babel-eslint": "^4.1.5",
-    "babelify": "~5.0.4",
-    "browserify": "~9.0.3",
-    "eslint": "^1.9.0",
-    "eslint-config-airbnb": "^1.0.0",
-    "eslint-plugin-react": "^3.8.0",
-    "gulp": "^3.8.11",
-    "gulp-autoprefixer": "^2.1.0",
-    "gulp-concat": "^2.4.3",
-    "gulp-minify-css": "^0.4.5",
-    "gulp-rename": "^1.2.0",
-    "gulp-ruby-sass": "^1.0.0-alpha.3",
-    "gulp-sourcemaps": "^1.3.0",
-    "gulp-uglify": "^1.1.0",
-    "gulp-util": "~3.0.4",
-    "gulp-watch": "~4.2.0",
-    "react-addons-perf": "^0.14.2",
-    "vinyl-buffer": "~1.0.0",
-    "vinyl-source-stream": "~1.1.0",
-    "watchify": "~2.4.0"
+    "aliasify": "^2.1.0",
+    "babel-eslint": "^8.0.1",
+    "babelify": "^7.3.0",
+    "browserify": "^14.4.0",
+    "electron": "1.8.1",
+    "eslint": "^4.8.0",
+    "eslint-config-standard": "^10.2.1",
+    "eslint-config-standard-jsx": "^4.0.2",
+    "eslint-formatter-pretty": "^1.3.0",
+    "eslint-plugin-compat": "^1.0.4",
+    "eslint-plugin-import": "^2.7.0",
+    "eslint-plugin-jsx-a11y": "^6.0.2",
+    "eslint-plugin-node": "^5.2.0",
+    "eslint-plugin-promise": "^3.5.0",
+    "eslint-plugin-react": "^7.4.0",
+    "eslint-plugin-standard": "^3.0.1",
+    "gulp": "^3.9.1",
+    "gulp-autoprefixer": "^4.0.0",
+    "gulp-concat": "^2.6.1",
+    "gulp-minify-css": "^1.2.4",
+    "gulp-rename": "^1.2.2",
+    "gulp-ruby-sass": "^2.1.1",
+    "gulp-sourcemaps": "^2.6.1",
+    "gulp-uglify": "^3.0.0",
+    "gulp-util": "^3.0.8",
+    "gulp-watch": "^4.3.11",
+    "react-addons-perf": "^15.4.2",
+    "vinyl-buffer": "^1.0.0",
+    "vinyl-source-stream": "^1.1.0",
+    "watchify": "^3.9.0"
   },
   "browser": {
     "babel/register": false
