@@ -1,32 +1,29 @@
-const React           = require('react');
-const Reflux          = require('reflux');
-const Menu            = require('./Menu.jsx');
-const interfaceStore  = require('../stores/interface');
+import React from 'react'
+import { connect } from 'react-redux'
+import Menu from './Menu.jsx'
 
-const ApplicationMenu = React.createClass({
-    mixins: [Reflux.ListenerMixin],
-    getInitialState() {
-        return {
-            menu: interfaceStore.getApplicationMenu(),
-        };
-    },
-    componentDidMount() {
-        this.listenTo(interfaceStore, this.oninterfaceStoreChange);
-    },
-    oninterfaceStoreChange() {
-        this.setState({
-            menu: interfaceStore.getApplicationMenu(),
-        });
-    },
-    render() {
-        if (!this.state.menu) {
-            return null;
-        }
-        return (
-            <div className="application-menu">
-                <Menu data={this.state.menu}/>
-            </div>
-        );
-    },
-});
-module.exports = ApplicationMenu;
+class ApplicationMenu extends React.Component {
+  constructor (props) {
+    super(props)
+    this.state = {}
+  }
+  render () {
+    if (!this.props.menu) {
+      return null
+    }
+    return (
+      <div className='application-menu'>
+        <Menu data={this.props.menu} />
+      </div>
+    )
+  }
+}
+
+function mapStateToProps (state) {
+  let menu = state.ui.get('menu')
+  return {
+    menu: menu
+  }
+}
+
+export default connect(mapStateToProps)(ApplicationMenu)
