@@ -1,19 +1,15 @@
-const React = require('react');
+import React from 'react'
 
-const Group = React.createClass({
-    propTypes: {
-        title: React.PropTypes.string,
-        children: React.PropTypes.element,
-    },
-    render() {
-        return (
-            <div className="group">
-                <fieldset>
-                    <legend>{this.props.title}</legend>
-                    {this.props.children}
-                </fieldset>
-            </div>
-        );
-    },
-});
-module.exports = Group;
+class Group extends React.Component {
+  render () {
+    return (
+      <div className='group'>
+        <fieldset>
+          <legend>{this.props.title}</legend>
+          {this.props.children}
+        </fieldset>
+      </div>
+    )
+  }
+}
+export default Group
