@@ -1,52 +1,52 @@
-const React = require('react');
+import React from 'react'
+import { acceleratorToText } from '../lib/accelerator'
 
-const Menu = React.createClass({
-    propTypes: {
-        data: React.PropTypes.object, // Menu Data
-    },
-    renderItem(item, key) {
-        if (item.type === 'separator') {
-            return (
-                <div key={key} className="menu-separator"></div>
-            );
-        }
+class Menu extends React.Component {
+  handleClick (item) {
+    if (item.click) {
+      item.click()
+    }
+  }
 
-        const classes = [
-            'menu-item',
-        ];
+  renderItem (item, key) {
+    if (item.type === 'separator') {
+      return (
+        <div key={key} className='menu-separator' />
+      )
+    }
 
-        let subMenu = null;
-        if (item.submenu) {
-            classes.push('parent');
-            subMenu = (
-                <Menu data={item.submenu}/>
-            );
-        }
-        let accelerator = null;
-        if (item.globalKeyObject) {
-            accelerator = item.globalKeyObject.text;
-        }
-        return (
-            <div key={key} className={classes.join(' ')} data-accelerator={accelerator} onClick={item.click}>
-                {item.label}
-                {subMenu}
-            </div>
-        );
-    },
+    const classes = [
+      'menu-item'
+    ]
 
-    render() {
-        if (!this.props.data) {
-            return null;
-        }
-        const menuItems = this.props.data.items.map((item, i) => {
-            return this.renderItem(item, i);
-        });
-        return (
-            <div className="menu">
-                {menuItems}
-            </div>
-        );
-    },
-});
+    let subMenu = null
+    if (item.submenu) {
+      classes.push('parent')
+      subMenu = (
+        <Menu data={item.submenu} />
+      )
+    }
+    return (
+      <div key={key} className={classes.join(' ')} data-accelerator={acceleratorToText(item.accelerator)} onClick={this.handleClick.bind(this, item)}>
+        {item.label}
+        {subMenu}
+      </div>
+    )
+  }
 
-module.exports = Menu;
+  render () {
+    if (!this.props.data) {
+      return null
+    }
+    const menuItems = this.props.data.items.map((item, i) => {
+      return this.renderItem(item, i)
+    })
+    return (
+      <div className='menu'>
+        {menuItems}
+      </div>
+    )
+  }
+}
+
+export default Menu
