@@ -1,72 +1,79 @@
-const React = require('react');
-const _ = require('lodash');
-const interfaceActions = require('../actions/interface');
+import React from 'react'
+import { connect } from 'react-redux'
+import { bindActionCreators } from 'redux'
+import Immutable from 'immutable'
+import { throttle } from 'lodash'
+import { setTreeSize } from '../actions/interface'
 
-const TreeHandle = React.createClass({
-    propTypes: {
-        index: React.PropTypes.number, // Column Index
-        size: React.PropTypes.number, // Column Size
-    },
+class TreeHandle extends React.Component {
+  componentWillMount () {
+    this.handleMouseMove = throttle(this.handleMouseMove, 30)
+  }
 
-    componentWillMount() {
-        this.handleMouseMove = _.throttle(this.handleMouseMove, 30);
-    },
+  componentDidMount () {
+    window.addEventListener('mouseup', this.handleMouseUp)
+  }
 
-    componentDidMount() {
-        window.addEventListener('mouseup', this.handleMouseUp);
-    },
+  componentWillUnmount () {
+    window.removeEventListener('mouseup', this.handleMouseMove)
+    window.removeEventListener('mousemove', this.handleMouseMove)
+  }
 
-    componentWillUnmount() {
-        window.removeEventListener('mouseup', this.handleMouseMove);
-        window.removeEventListener('mousemove', this.handleMouseMove);
-    },
+  handleMouseMove = (event) => {
+    const diff = event.clientX - this.startPos
+    const newSize = this.startSize + diff
+    this.props.setTreeSize(this.props.index, newSize)
+  }
 
-    handleMouseMove(event) {
-        const diff = event.clientX - this.startPos;
-        const newSize = this.startSize + diff;
-        interfaceActions.setTreeSize(this.props.index, newSize);
-    },
+  handleMouseUp = (event) => {
+    window.removeEventListener('mousemove', this.handleMouseMove, false)
+  }
 
-    handleMouseUp(event) {
-        window.removeEventListener('mousemove', this.handleMouseMove, false);
-    },
+  handleMouseDown = (event) => {
+    this.startPos = event.clientX
+    this.startSize = this.props.size
 
-    handleMouseDown(event) {
-        this.startPos = event.clientX;
-        this.startSize = this.props.size;
+    window.addEventListener('mousemove', this.handleMouseMove, false)
+  }
 
-        window.addEventListener('mousemove', this.handleMouseMove, false);
-    },
+  render () {
+    return (
+      <div ref='handle' className='tree-handle' onMouseDown={this.handleMouseDown} />
+    )
+  }
+}
 
-    render() {
-        return (
-            <div ref="handle" className="tree-handle" onMouseDown={this.handleMouseDown}></div>
-        );
-    },
-});
+function mapStateToProps (state) {
+  return {}
+}
 
-const TreeHeader = React.createClass({
-    propTypes: {
-        columns: React.PropTypes.arrayOf(React.PropTypes.string),
-        sizes: React.PropTypes.arrayOf(React.PropTypes.number),
-    },
-    shouldComponentUpdate(nextProps) {
-        return this.props.columns !== nextProps.columns || this.props.sizes !== nextProps.sizes;
-    },
-    render() {
-        return (
-            <div className="tree-header">
-                {this.props.columns.map((col, i) => {
-                    const style = {width: this.props.sizes[i]+'px'};
-                    return (
-                        <div key={i} className="tree-header-col" style={style}>
-                            {col}
-                            <TreeHandle index={i} size={this.props.sizes[i]} />
-                        </div>
-                    );
-                })}
+function mapDispatchToProps (dispatch) {
+  return bindActionCreators({setTreeSize}, dispatch)
+}
+
+const ConnectedTreeHandle = connect(mapStateToProps, mapDispatchToProps)(TreeHandle)
+
+class TreeHeader extends React.Component {
+  shouldComponentUpdate (nextProps) {
+    return this.props.columns !== nextProps.columns ||
+      !Immutable.is(this.props.sizes, nextProps.sizes)
+  }
+
+  render () {
+    return (
+      <div className='tree-header'>
+        {this.props.columns.map((col, i) => {
+          const style = {width: this.props.sizes.get(i) + 'px'}
+          return (
+            <div key={i} className='tree-header-col' style={style}>
+              {col}
+              <ConnectedTreeHandle index={i} size={this.props.sizes.get(i)} />
             </div>
-        );
-    },
-});
-module.exports = TreeHeader;
+          )
+        })}
+      </div>
+    )
+  }
+}
+
+export default TreeHeader
