@@ -1,125 +1,132 @@
-const xml2js = require('xml2js');
-const path   = require('path');
-const fs     = require('fs');
-const _      = require('lodash');
+import xml2js from 'xml2js'
+import { extname, basename } from 'path'
+import { exists, readFile, writeFile } from 'fs'
+import { padStart } from 'lodash'
+import { saveAs } from 'file-saver'
+import { remote } from 'electron'
+const { dialog } = remote
 
-const workspaceActions = require('../actions/workspace');
-
-function loadProfile(profilePath, callback) {
-    console.log("loading profile");
-    var profileExt = path.extname(profilePath);
-    fs.readFile(profilePath, function(err, data) {
-        if (err) {
-            console.log(err);
-            callback(null);
-        } else {
-            if (profileExt === '.json') {
-                try {
-                    var json = JSON.parse(data);
-                } catch (e) {
-                    console.error('There was an error parsing '+profilePath);
-                    console.log(e);
-                    return;
-                }
-                callback(json);
-            } else if (profileExt === '.xml') {
-                var parser = new xml2js.Parser({
-                    tagNameProcessors: [function(name){
-                        if (name == 'directory')    return 'directories';
-                        else if (name == 'texture') return 'textures';
-                        return name;
-                    }],
-                    attrNameProcessors: [],
-                    attrkey: 'data',
-                });
-                parser.parseString(data, function (err, result) {
-                    if (err) {
-                        console.error(err);
-                        return;
-                    }
-
-                    const relativeToAbsolute = function(item, parentAddress) {
-                        if (!('address' in item.data)) {
-                            item.data.address = '0x00000000';
-                        }
-
-                        item.data.address = '0x'+_.padLeft((parseInt(item.data.address, 16) + parseInt(parentAddress, 16)).toString(16), 8, 0).toUpperCase();
-                        if (item.textures) {
-                            item.textures.forEach((texture) => {
-                                relativeToAbsolute(texture, item.data.address);
-                            });
-                        }
-                        if (item.directories) {
-                            item.directories.forEach((directory) => {
-                                relativeToAbsolute(directory, item.data.address);
-                            });
-                        }
-                    };
-
-                    relativeToAbsolute(result.root, '0x0');
-                    console.log(result.root);
-
-                    console.log(JSON.stringify(result.root, null, 4));
-
-                    callback(result.root);
-                });
-            }
-        }
-    });
-}
-
-function addToRecent(file) {
-    var recent = localStorage.getItem('recent_files');
-    if (recent !== null) {
-        recent = JSON.parse(recent);
+export function loadProfile (profilePath, callback) {
+  var profileExt = extname(profilePath)
+  readFile(profilePath, function (err, data) {
+    if (err) {
+      console.error(err)
+      callback(null)
     } else {
-        recent = [];
-    }
-    if (recent.indexOf(file) >= 0) {
-        recent.splice(recent.indexOf(file),1);
-    }
-    recent.unshift(file);
-    recent = recent.slice(0,5);
-    localStorage.setItem('recent_files', JSON.stringify(recent));
-}
-
-function clearRecent() {
-    localStorage.setItem('recent_files', JSON.stringify([]));
-}
-
-function openFile(filePath, callback) {
-    fs.exists(filePath, (exists) => {
-        if (exists) {
-            fs.readFile(filePath, (err, data) => {
-                if (err) throw err;
-                addToRecent(filePath);
-
-                workspaceActions.createWorkspace({
-                    data: data,
-                    path: filePath,
-                    name: path.basename(filePath),
-                });
-                if (callback) {
-                    callback();
-                }
-            });
+      if (profileExt === '.json') {
+        try {
+          var json = JSON.parse(data)
+        } catch (e) {
+          console.error('There was an error parsing ' + profilePath)
+          console.error(e)
+          return
         }
-    });
+        callback(json)
+      } else if (profileExt === '.xml') {
+        var parser = new xml2js.Parser({
+          tagNameProcessors: [function (name) {
+            if (name === 'directory') return 'directories'
+            else if (name === 'texture') return 'textures'
+            return name
+          }],
+          attrNameProcessors: [],
+          mergeAttrs: true,
+          explicitArray: false
+        })
+        parser.parseString(data, function (err, result) {
+          if (err) {
+            console.error(err)
+            return
+          }
+
+          const relativeToAbsolute = function (item, parentAddress = '0x0') {
+            let address = '0x00000000'
+            if ('address' in item) { // Root does not have an address
+              item.address = '0x' + padStart((parseInt(item.address, 16) + parseInt(parentAddress, 16)).toString(16), 8, 0).toUpperCase()
+              address = item.address
+            }
+            if ('palette' in item) {
+              item.palette = '0x' + padStart((parseInt(item.palette, 16) + parseInt(parentAddress, 16)).toString(16), 8, 0).toUpperCase()
+            }
+            if (item.textures) {
+              if (!Array.isArray(item.textures)) {
+                item.textures = [item.textures]
+              }
+              item.textures.forEach((texture) => {
+                relativeToAbsolute(texture, address)
+              })
+            }
+            if (item.directories) {
+              if (!Array.isArray(item.directories)) {
+                item.directories = [item.directories]
+              }
+              item.directories.forEach((directory) => {
+                relativeToAbsolute(directory, address)
+              })
+            }
+          }
+          relativeToAbsolute(result.root)
+          callback(result.root)
+        })
+      }
+    }
+  })
 }
 
-function saveFile(filePath, fileData, callback) {
-    fs.writeFile(filePath, fileData, {}, (err) => {
-        if (err) throw err;
-        if (callback) {
-            callback(err);
-        }
-    });
+// We handle File & path
+export function openFile (input, callback) {
+  if (input instanceof File) {
+    let file = input
+    const reader = new FileReader()
+    event.preventDefault()
+    reader.onloadend = (e) => {
+      callback({
+        data: Buffer.from(e.target.result),
+        path: file.path || file.name,
+        name: file.name
+      })
+    }
+    reader.readAsArrayBuffer(file)
+  } else if (typeof input === 'string') {
+    let path = input
+    exists(path, (exists) => {
+      if (exists) {
+        readFile(path, (err, data) => {
+          if (err) throw err
+          if (callback) {
+            callback({
+              data: data,
+              path: path,
+              name: basename(path)
+            })
+          }
+        })
+      }
+    })
+  }
 }
 
-module.exports = {
-    addToRecent,
-    clearRecent,
-    loadProfile,
-    openFile,
-    saveFile,
-};
+// Used to save files to the OS
+export function saveFileAs (filePath, fileData, callback) {
+  dialog.showSaveDialog(remote.getCurrentWindow(), {
+    title: 'Save File',
+    defaultPath: filePath
+  }, (filePath) => {
+    if (filePath) {
+      if (process.browser) {
+        let blob = new Blob([fileData])
+        saveAs(blob, filePath)
+      } else {
+        writeFile(filePath, fileData, {}, (err) => {
+          if (err) throw err
+          if (callback) {
+            callback(err)
+          }
+        })
+      }
+    } else {
+      let err = new Error('No file name provided')
+      callback(err)
+    }
+  })
+}
