@@ -1,18 +1,24 @@
-var fileOpener = document.createElement('input');
-fileOpener.type = 'file';
-fileOpener.id = 'fileOpener';
-fileOpener.style.display = 'none';
+var fileOpener = document.createElement('input')
+fileOpener.type = 'file'
+fileOpener.id = 'fileOpener'
+fileOpener.style.display = 'none'
 
-document.body.insertBefore(fileOpener, document.body.firstChild);
+document.body.insertBefore(fileOpener, document.body.firstChild)
 
-module.exports = {
-    showOpenDialog(w, options, callback) {
-        var fileOpener = document.getElementById('fileOpener');
-        fileOpener.onchange = callback;
-        fileOpener.click();
-    },
-    showSaveDialog(w, options, callback) {
-        var fileName = prompt("File Name", "");
-        callback(fileName);
-    }
-};
+export function showOpenDialog (w, options, callback) {
+  var fileOpener = document.getElementById('fileOpener')
+  fileOpener.onchange = function (event) {
+    callback(this.files)
+  }
+  fileOpener.click()
+}
+
+export function showSaveDialog (w, options, callback) {
+  var fileName = prompt('File Name', options.defaultPath)
+  callback(fileName)
+}
+
+export default {
+  showOpenDialog,
+  showSaveDialog
+}
