@@ -1,38 +1,51 @@
-require("babel/register");
-var textureManipulator = require('../textureManipulator');
+require('babel-register')
+var textureManipulator = require('../textureManipulator')
 
-process.title = 'TE.js Worker';
+function work (msg, callback) {
+  var input = msg.input
+  if (input.type !== 'generatePNGBuffer') {
+    return
+  }
+  var palette
+  if (input.palette) {
+    palette = Buffer.from(input.palette)
+  }
 
-var work = function(msg, callback){
-    var input = msg.input;
-    if (input.type != 'generateTexture') {
-        return;
-    }
-    var palette;
-    if (input.palette) {
-        palette = new Buffer(input.palette);
-    }
+  var img = textureManipulator.generateTexture(Buffer.from(input.data), input.format, input.width, palette)
+  img.toPNGBuffer((buffer) => {
+    // TODO: handle errors...?
+    callback(null, {
+      id: msg.id,
+      type: msg.type,
+      output: {
+        img: img,
+        buffer: buffer
+      }
+    })
+  })
+}
 
-    var img = textureManipulator.generateTexture(new Buffer(input.data), input.format, input.width, palette);
-    var out = {
-        id: msg.id,
-        type: msg.type,
-        output: {img: img}
-    };
-
-    if (callback)
-        callback(out);
-    else
-        process.send(out);
-};
-process.on('message', work);
-
-module.exports = { //Process shim for Browser Mode
-    _callBack: undefined,
-    on: function(type, callback) {
-        this._callBack = callback;
-    },
-    send: function(data) {
-        work(data, this._callBack);
-    }
-};
+if (process.browser) {
+  module.exports = function (self) {
+    self.addEventListener('message', function (ev) {
+      work(ev.data, (err, msg) => {
+        if (err) {
+          self.postMessage(err)
+        } else {
+          self.postMessage(msg)
+        }
+      })
+    })
+  }
+} else {
+  process.title = 'TE.js Worker'
+  process.on('message', function (msg) {
+    work(msg, (err, msg) => {
+      if (err) {
+        process.send(err)
+      } else {
+        process.send(msg)
+      }
+    })
+  })
+}
