@@ -1,28 +1,42 @@
-var app = require('app');
-var path = require('path');
-var fs = require('fs');
-var BrowserWindow = require('browser-window');
-argv = require('minimist')(process.argv.slice(1));
+const { app, BrowserWindow } = require('electron')
+const windowStateKeeper = require('electron-window-state')
+global.argv = require('minimist')(process.argv.slice(1))
+const path = require('path')
 
-require('crash-reporter').start();
+var mainWindow = null
 
-var mainWindow = null;
+app.on('window-all-closed', function () {
+  if (process.platform !== 'darwin') {
+    app.quit()
+  }
+})
 
-app.on('window-all-closed', function() {
-    if (process.platform != 'darwin')
-        app.quit();
-});
+app.on('ready', function () {
+  let mainWindowState = windowStateKeeper({
+    defaultWidth: 1280,
+    defaultHeight: 600
+  })
 
-app.on('ready', function() {
-    mainWindow = new BrowserWindow({width: 800, height: 600, icon: path.join(__dirname, 'icon.png')});
+  mainWindow = new BrowserWindow({
+    x: mainWindowState.x,
+    y: mainWindowState.y,
+    width: mainWindowState.width,
+    height: mainWindowState.height,
+    icon: path.join(__dirname, 'icon.png'),
+    backgroundColor: '#262626'
+  })
 
-    mainWindow.loadUrl('file://' + __dirname + '/index.html');
+  mainWindowState.manage(mainWindow)
 
-    if (argv._.indexOf('debug') > -1) {
-        mainWindow.toggleDevTools();
-    }
+  mainWindow.loadURL('file://' + __dirname + '/index.html')
 
-    mainWindow.on('closed', function() {
-        mainWindow = null;
-    });
-});
+  if (global.argv._.indexOf('debug') > -1) {
+    mainWindow.openDevTools({
+      mode: 'bottom'
+    })
+  }
+
+  mainWindow.on('closed', function () {
+    mainWindow = null
+  })
+})
